% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_Ref.R
\name{OR_Ref}
\alias{OR_Ref}
\title{Reference Analysis of the 2x2 table}
\usage{
OR_Ref(n1, n2, n3, n4, conf.int = TRUE, conf.level = 0.95,
  post.sample = TRUE, sampling.depth = 4, num.samples = 1000)
}
\arguments{
\item{n1}{Upper left element of the table}

\item{n2}{Upper right element of the table}

\item{n3}{Lower left element of the table}

\item{n4}{Lower right element of the table}

\item{conf.int}{Should a credible interval be computed using numerical integration?}

\item{conf.level}{confidence level for the returned credible interval. NOTE: Equal tailed}

\item{post.sample}{Should Posterior sampling be conducted?}

\item{sampling.depth}{Depth of sampling.  1 for theta only, 2 to add eta1, 3 to add eta3, 4 to add eta2}

\item{num.samples}{Number of Posterior Samples}
}
\description{
Perform Bayesian analysis on the 2x2 table using a reference prior
}
\examples{
n1=5
n2=0
n3=8
n4=15

res <- OR_Ref(n1,n2,n3,n4,conf.int = FALSE,num.samples = 500)
}
